+++ perf_benchmark_test.go
@@ -293,20 +293,22 @@ func validatePerfPacketBytes(wl perfWorkload, pkt []byte) error {
 }
 
 func perfMusicFrame(start, n, channels int) []float64 {
+	// Products are rounded explicitly so the fixture is identical on every
+	// architecture (no fused multiply-add; see docs/DEVELOPER.md).
 	out := make([]float64, n*channels)
 	for i := 0; i < n; i++ {
-		t := float64(start+i) / perfSampleRate
-		env := 0.55 + 0.08*math.Sin(2*math.Pi*3.1*t)
-		left := env * (0.28*math.Sin(2*math.Pi*440*t) +
-			0.18*math.Sin(2*math.Pi*1760*t+0.2) +
-			0.10*math.Sin(2*math.Pi*6200*t+0.7) +
-			0.05*math.Sin(2*math.Pi*15500*t+1.1))
+		t := float64(float64(start+i) / perfSampleRate)
+		env := 0.55 + float64(0.08*math.Sin(float64(2*math.Pi*3.1*t)))
+		left := float64(env * (float64(0.28*math.Sin(float64(2*math.Pi*440*t))) +
+			float64(0.18*math.Sin(float64(2*math.Pi*1760*t)+0.2)) +
+			float64(0.10*math.Sin(float64(2*math.Pi*6200*t)+0.7)) +
+			float64(0.05*math.Sin(float64(2*math.Pi*15500*t)+1.1))))
 		out[i*channels] = left
 		if channels == 2 {
-			right := env * (0.25*math.Sin(2*math.Pi*554.37*t+0.3) +
-				0.16*math.Sin(2*math.Pi*2217.46*t+0.5) +
-				0.09*math.Sin(2*math.Pi*7600*t+0.9) +
-				0.05*math.Sin(2*math.Pi*14200*t+1.4))
+			right := float64(env * (float64(0.25*math.Sin(float64(2*math.Pi*554.37*t)+0.3)) +
+				float64(0.16*math.Sin(float64(2*math.Pi*2217.46*t)+0.5)) +
+				float64(0.09*math.Sin(float64(2*math.Pi*7600*t)+0.9)) +
+				float64(0.05*math.Sin(float64(2*math.Pi*14200*t)+1.4))))
 			out[i*channels+1] = right
 		}
 	}
